Added split_allocation
The ability to split existing allocations as unobtrusively as possible is incredibly useful, and is necessary for things such as allocating a stack with an overflow guard (by allocating n+1 bytes then calling split(alloc, 1) -> (guardpage, stackpages).

At least I think it will be. It adds flexibility and for little cost.