Added VMEM allocation strategies, to encourage the allocator to place certain allocations in sensible locations.
For example, the stack can be placed at the end of the relevant part of
VMEM, and if multiple stacks are present they can be spread out. This is
useful for user tasks which may want to request a stack expansion at
random points (which would otherwise be impossible).

Allocating extra kernel stacks in a non offset-mapped area is not an
issue, for two reasons: 1. page tables are on the heap (since they're
behind an Arc if you use LockedPageAllocator/PagingContext), and 2.
the kernel heap and task stacks are unlikely to use the same memory,
unless someone invents a computer with 512GB of ram and somehow coerces
the kernel to use memory near or above the 512GB limit (if we get close
to this, future me needs to tell the kernel to reserve physmem >512GB
for user applications, stacks, etc, though I doubt it'll happen as
general relativity makes it near-impossible for that amount of RAM to be
useful (at current memory densities) unless you're using a NUMA
architecture, which this kernel isn't designed for anyway.