Made a first attempt at structuring how "descriptors" will work.
This is still a WIP and may be subject to change.

Descriptors are used for managing lots of things, from files to
processes. As an example, in a process, things like memory mappings are
stored in the B-slot (with B-references being held by threads). When the
last thread exits, B-slot values such as memory mappings are freed, but
the rest will remain until all A-references are cleared (i.e. the parent
or siblings are done with the process)