Stacks are now stored alongside their tasks instead of being leaked.
They have to be dropped after another task has executed, however, as
their stack is still in use by the scheduler when switching immediately
away from them.

I would just attach the stack to the task, and add a field for "current
task stack", but unfortunately LockedPageAllocator and AllocatedStack
are generic over PageFrameAllocator. That was probably my billion dollar
mistake. :(